Add allowed() helper function to use in templates
This helper function gets the needed privilege for a resource and checks
the logged in user permissions against that privilege, if the user is
allowed to view that resource [a link, picture, whatever], it returns
true.
Inside the template, you only need to put that resource inside an if
condition, using this allowed() helper function.