Refactor test: add user delete in separate method
By using haveInDatabase, you can add your desired user in the DB first
and then write your expectations.
Remember: each method in a Cest file is run separately than the others
and the DB is reset for every method. That means if you add user
'ghelgheli' in one method and you expect to see that user in another
method to delete it, you get screwed! It won't be there!